pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "earthquake-analysis"
version = "0.1.0"
description = "Analyze global earthquake patterns and real-world impact using USGS and NOAA/NCEI data."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31",
    "pandas>=2.0",
    "numpy>=1.26",
    "python-dateutil>=2.8",
    "streamlit>=1.35",
    "plotly>=5.20",
]

[project.optional-dependencies]
dev = ["pytest>=8", "black", "ruff"]

[tool.setuptools.packages.find]
where = ["."]
include = ["earthquake_analysis*"]
